Rename namespace Column into TableColumn to avoid confusion with Columns::addColumn
