+++ src/TableColumn/Template.php
@@ -1,14 +1,24 @@
 <?php
 
-namespace atk4\ui\Column;
+namespace atk4\ui\TableColumn;
 
 /**
  * Implements Column helper for grid.
  */
 class Template extends Generic
 {
+    /**
+     * User-defined template for this Column
+     *
+     * @var string
+     */
     public $template = null;
 
+    /**
+     * call new TableColumn\Template('{$name} {$surname}');
+     *
+     * @param string $template Template with {$tags}
+     */
     public function __construct($template)
     {
         $this->template = $template;
